Probe Windows path aliases through the filesystem API
cmd.exe quoting is not stable through Go's argument escaping. Use GetShortPathNameW directly in the Windows-only regression helper and keep a compile-time non-Windows skip twin.

Constraint: The regression must derive aliases from the active volume without hardcoded runner paths or shell parsing.

Rejected: Post-process cmd's malformed quoted output | that would couple the shield to one parser accident.

Confidence: high

Scope-risk: narrow

Tested: Two focused runs reproduced the cmd quoting failure independently of production behavior.

Not-tested: Syscall-backed helper pending restarted reviews and focused rerun.
